project_3/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0
scipy>=1.10.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
albumentations>=1.3.0
clean-fid>=0.1.35
torchmetrics>=0.11.0
wandb>=0.15.0
pyyaml>=6.0
einops>=0.6.0
accelerate>=0.20.0
diffusers>=0.21.0
transformers>=4.30.0
kornia>=0.7.0